Remove improper unit tests
